requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-functions-durable
psycopg2
pandas
langchain_core
langchain
langchain_openai
langchain_azure_dynamic_sessions
langchain_experimental
langchain_community
azure-identity
azure-storage-blob
